requirements.txt
dpkt
colorama
python-geoip-geolite2